change namespace of scripts
change in namespace
