+++ Assets/Script/Runtime/SpawnManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83dc1cdc535ca8b42bad555ac4958b25
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
